# docs/public/openapi.yaml
openapi: 3.1.0
info:
  title: BetterStatusPage admin API
  version: 1.0.0
  description: |
    The part of the admin API meant for scripts, CI/CD pipelines and infrastructure tooling. It is the same API the
    admin panel uses, limited to what an API token may call; see https://docs.betterstatuspage.dev/api/.

    **Authentication.** Send an API token (Administration → API tokens) as `Authorization: Bearer bsp_...`. A token
    carries permissions per part of the application: `monitors`, `channels`, `incidents`, `maintenance`,
    `subscribers`, `appearance` (layout and branding) and `reports`, `audit`, `system` (read only), each `read` or
    `write` (write includes read). The separate `vault:use` permission lets a token attach vault secrets to monitors
    and channels; without it a token cannot add, change or move a vault reference. Each operation below names the
    permission it needs in `x-scope`. A token expires after 90 days unless it was created without an expiry.

    **Secrets.** Passwords, client secrets, credential headers, Slack, Discord and Teams webhook URLs and bot tokens in a `config` are never
    returned: they read back as `••••••••`. Send that value back unchanged to keep the stored secret, or send a new
    value (text) to replace it. A masked value that does not match the stored secret, or one that looks like a
    half-edited mask, is refused with `400`. A kept secret must stay where it is sent: if the `url` (or the OAuth2
...

...

...
servers:
...
    variables:
...
security:
  - bearerAuth: []
...
paths:
  /monitors:
    get:
      tags: [Monitors]
      summary: List monitors
      x-scope: monitors:read
      responses:
        '200':
          description: Every monitor, with secrets masked.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Monitor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Monitors]
      summary: Create a monitor
      x-scope: monitors:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonitorCreate' }
      responses:
        '200':
          description: The created monitor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /monitors/test:
    post:
      tags: [Monitors]
      summary: Test a monitor configuration without saving it
      description: Runs the check once and reports each step. Supported for `https`, `ping`, `dns`, database and `docker` monitors.
      x-scope: monitors:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, config]
              properties:
                type: { $ref: '#/components/schemas/MonitorType' }
                config: { type: object }
                timeoutMs: { type: integer, default: 10000, description: 'Milliseconds, used from 500 to 60000: a value outside that is clamped.' }
                monitorId:
                  type: integer
                  description: The saved monitor being edited, so a masked secret in `config` is tested as the stored one.
      responses:
        '200':
          description: The result of the test.
          content:
            application/json:
              schema:
                type: object
                properties:
                  overall: { type: string, enum: [ok, error] }
                  totalMs: { type: integer }
                  steps:
                    type: array
                    items:
                      type: object
                      properties:
                        label: { type: string }
                        status: { type: string, enum: [ok, error, info] }
                        detail: { type: string }
                        durationMs: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /monitors/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Monitors]
      summary: Get a monitor
      x-scope: monitors:read
      responses:
        '200':
          description: The monitor, with secrets masked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Monitors]
      summary: Update a monitor
      description: Only the fields you send change. `config` replaces the whole configuration, and a change of `type` needs it.
      x-scope: monitors:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonitorUpdate' }
      responses:
        '200':
          description: The updated monitor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Monitors]
      summary: Delete a monitor
      x-scope: monitors:write
      responses:
        '204': { description: Deleted (also when it did not exist). }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /monitors/{id}/check-now:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Monitors]
      summary: Run a check immediately
      x-scope: monitors:write
      responses:
        '200':
          description: The monitor after the check.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monitor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /monitors/{id}/dependencies:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Monitors]
      summary: List the monitors this monitor depends on
      x-scope: monitors:read
      responses:
        '200':
          description: Ids of the monitors it depends on.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dependencies' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Monitors]
      summary: Replace the monitors this monitor depends on
      description: Unknown ids are ignored. A dependency cycle is refused.
      x-scope: monitors:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Dependencies' }
      responses:
        '200':
          description: Saved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /incidents/{id}/monitors:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Incidents]
      summary: Set the monitors an incident affects
      x-scope: incidents:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [monitorIds]
              properties:
                monitorIds:
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: The incident with its monitors.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /layout/monitors:
    get:
      tags: [Layout]
      summary: List monitors by name and type, for placing them in the layout
      x-scope: appearance:read
      responses:
        '200':
          description: Id, name and type of every monitor; no configuration.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    name: { type: string }
                    type: { $ref: '#/components/schemas/MonitorType' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API token created in Administration → API tokens.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: The request is not valid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, unknown, expired or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The token does not have the permission this operation needs (the message names it), or the user's role may not do it.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The key is already used by another object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Key:
      type: string
      pattern: '^[a-z0-9][a-z0-9_-]{0,63}$'
      description: 'Unique, lowercase letters, digits, `-` and `_`. A technical value: it can be given when an object is created, and never changes after that.'
    Dependencies:
      type: object
      required: [dependsOnIds]
      properties:
        dependsOnIds:
          type: array
          items: { type: integer }
    MonitorType:
      type: string
      enum: [https, ping, dns, sqlserver, postgresql, mysql, mongodb, docker, webhook]
    MonitorTag:
      type: object
      required: [label, color]
      properties:
        label: { type: string }
        color: { type: string }
    VaultRef:
      type: object
      description: A secret read from a vault instead of being stored in the configuration.
      required: [vaultId, secretId]
      properties:
        vaultId: { type: integer }
        secretId: { type: integer }
        fieldMapping:
          type: object
          additionalProperties: { type: string }
    MonitorConfig:
      description: |
        The configuration for the monitor's `type`. A `webhook` monitor takes an empty object. Secrets are masked in
        responses (see the API description).
      oneOf:
        - $ref: '#/components/schemas/HttpsConfig'
        - $ref: '#/components/schemas/PingConfig'
        - $ref: '#/components/schemas/DnsConfig'
        - $ref: '#/components/schemas/DatabaseConfig'
        - $ref: '#/components/schemas/DockerConfig'
        - type: object
          maxProperties: 0
    HttpsConfig:
      type: object
      required: [url]
      properties:
        url: { type: string }
        method: { type: string, enum: [GET, POST, HEAD], default: GET }
        expectedStatus: { type: integer, default: 200 }
        keyword: { type: string }
        headers:
          type: object
          description: Values of credential headers are masked. A header is one when its name contains auth, token, secret, key, pass, pwd, credential, cookie, session or signature.
          additionalProperties: { type: string }
        body: { type: string }
        auth:
          type: object
          properties:
            type: { type: string, enum: [none, basic, oauth2, cas] }
            basic:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                vault: { $ref: '#/components/schemas/VaultRef' }
            oauth2:
              type: object
              properties:
                tokenUrl: { type: string }
                clientId: { type: string }
                clientSecret: { type: string }
                scope: { type: string }
                vault: { $ref: '#/components/schemas/VaultRef' }
            cas:
              type: object
              properties:
                casServerUrl: { type: string }
                username: { type: string }
                password: { type: string }
                vault: { $ref: '#/components/schemas/VaultRef' }
        certExpiry:
          type: object
          properties:
            enabled: { type: boolean }
            warnDays: { type: integer, minimum: 1, maximum: 365 }
    PingConfig:
      type: object
      required: [host, mode]
      properties:
        host: { type: string }
        mode: { type: string, enum: [tcp, icmp] }
        port: { type: integer }
    DnsConfig:
      type: object
      required: [hostname, recordType]
      properties:
        hostname: { type: string }
        recordType: { type: string, enum: [A, AAAA, MX, CNAME, TXT] }
        expectedValue: { type: string }
        resolver: { type: string }
    DatabaseConfig:
      type: object
      description: Used by the `sqlserver`, `postgresql`, `mysql` and `mongodb` monitor types.
      required: [query]
      properties:
        host: { type: string }
        port: { type: integer }
        database: { type: string }
        user: { type: string }
        password: { type: string }
        query: { type: string, description: A SQL statement, or a JSON command for MongoDB. }
        expectedResult: { type: string }
        mode: { type: string, enum: [fields, connectionString], default: fields }
        vault: { $ref: '#/components/schemas/VaultRef' }
    DockerConfig:
      type: object
      required: [endpoint, container]
      properties:
        endpoint: { type: string, description: 'unix:///var/run/docker.sock, npipe:////./pipe/docker_engine or http(s)://host:port' }
        container: { type: string, description: Container name or id. }
    Monitor:
      type: object
      properties:
        id: { type: integer }
        key: { $ref: '#/components/schemas/Key' }
        name: { type: string }
        type: { $ref: '#/components/schemas/MonitorType' }
        intervalSecs: { type: integer }
        timeoutMs: { type: integer }
        retries: { type: integer }
        failureThreshold: { type: integer }
        recoveryThreshold: { type: integer }
        config: { $ref: '#/components/schemas/MonitorConfig' }
        tags:
          type: array
          items: { $ref: '#/components/schemas/MonitorTag' }
        currentStatus: { type: string, enum: [up, down, degraded, pending, affected] }
        lastCheckedAt: { type: [integer, 'null'] }
        webhookToken: { type: [string, 'null'], description: The heartbeat token of a `webhook` monitor. }
        createdAt: { type: integer }
        updatedAt: { type: integer }
    MonitorCreate:
      type: object
      required: [name, type]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        key: { $ref: '#/components/schemas/Key' }
        type: { $ref: '#/components/schemas/MonitorType' }
        intervalSecs: { type: integer, minimum: 10, maximum: 86400, default: 60 }
        timeoutMs: { type: integer, minimum: 1000, maximum: 300000, default: 10000 }
        retries: { type: integer, minimum: 1, maximum: 10, default: 1 }
        failureThreshold:
          type: integer
          default: 1
          description: Consecutive failed checks before an alert, 1 to 20. A value outside that is clamped, not refused.
        recoveryThreshold:
          type: integer
          default: 1
          description: Consecutive successful checks before a recovery, 1 to 20. A value outside that is clamped, not refused.
        config: { $ref: '#/components/schemas/MonitorConfig' }
        tags:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/MonitorTag' }
    MonitorUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        type: { $ref: '#/components/schemas/MonitorType' }
        intervalSecs: { type: integer, minimum: 10, maximum: 86400 }
        timeoutMs: { type: integer, minimum: 1000, maximum: 300000 }
        retries: { type: integer, minimum: 1, maximum: 10 }
        failureThreshold: { type: integer, description: '1 to 20; a value outside that is clamped.' }
        recoveryThreshold: { type: integer, description: '1 to 20; a value outside that is clamped.' }
        config: { $ref: '#/components/schemas/MonitorConfig' }
        tags:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/MonitorTag' }
    IncidentStatus:
      type: string
      enum: [investigating, identified, monitoring, resolved]
    IncidentImpact:
      type: string
      enum: [minor, major, critical]
    Incident:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        title: { type: string }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        impact: { $ref: '#/components/schemas/IncidentImpact' }
        startedAt: { type: integer }
        resolvedAt: { type: [integer, 'null'] }
        createdAt: { type: integer }
        updatedAt: { type: integer }
...